refactor routes path and imports
