+++ src/api/v1/comms/id.js
@@ -1,11 +1,11 @@
 import Boom from '@hapi/boom'
 import Joi from 'joi'
 
-import { NotFoundError } from '../../errors/not-found-error.js'
-import { httpStatusResult } from '../../schemas/common/response.js'
-import { successModel } from '../../schemas/comms/responses/id.js'
-import { failAction } from '../../api/common/helpers/fail-action.js'
-import { getCommsEventById } from '../../repos/comms/comms-message.js'
+import { NotFoundError } from '../../../errors/not-found-error.js'
+import { httpStatusResult } from '../../../schemas/common/response.js'
+import { successModel } from '../../../schemas/comms/responses/id.js'
+import { failAction } from '../../../api/common/helpers/fail-action.js'
+import { getCommsEventById } from '../../../repos/comms/comms-message.js'
 
 export default [{
   method: 'GET',
